move imported code to /external
